.claude/monitor-web/requirements.txt
Django>=5.0,<6.0
PyYAML>=6.0
psutil>=5.9
gitpython>=3.1
requests>=2.31
python-dateutil>=2.8
tabulate>=0.9